--- src/abi/FollowNFT.ts
@@ -1,413 +1,0 @@
-export const FollowNFT = [
-    {
-      inputs: [{ internalType: 'address', name: 'hub', type: 'address' }],
-      stateMutability: 'nonpayable',
-      type: 'constructor'
-    },
-    { inputs: [], name: 'BlockNumberInvalid', type: 'error' },
-    { inputs: [], name: 'InitParamsInvalid', type: 'error' },
-    { inputs: [], name: 'Initialized', type: 'error' },
-    { inputs: [], name: 'NotHub', type: 'error' },
-    { inputs: [], name: 'NotOwnerOrApproved', type: 'error' },
-    { inputs: [], name: 'SignatureExpired', type: 'error' },
-    { inputs: [], name: 'SignatureInvalid', type: 'error' },
-    { inputs: [], name: 'TokenDoesNotExist', type: 'error' },
-    { inputs: [], name: 'ZeroSpender', type: 'error' },
-    {
-      anonymous: false,
-      inputs: [
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'owner',
-          type: 'address'
-        },
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'approved',
-          type: 'address'
-        },
-        {
-          indexed: true,
-          internalType: 'uint256',
-          name: 'tokenId',
-          type: 'uint256'
-        }
-      ],
-      name: 'Approval',
-      type: 'event'
-    },
-    {
-      anonymous: false,
-      inputs: [
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'owner',
-          type: 'address'
-        },
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'operator',
-          type: 'address'
-        },
-        {
-          indexed: false,
-          internalType: 'bool',
-          name: 'approved',
-          type: 'bool'
-        }
-      ],
-      name: 'ApprovalForAll',
-      type: 'event'
-    },
-    {
-      anonymous: false,
-      inputs: [
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'from',
-          type: 'address'
-        },
-        {
-          indexed: true,
-          internalType: 'address',
-          name: 'to',
-          type: 'address'
-        },
-        {
-          indexed: true,
-          internalType: 'uint256',
-          name: 'tokenId',
-          type: 'uint256'
-        }
-      ],
-      name: 'Transfer',
-      type: 'event'
-    },
-    {
-      inputs: [],
-      name: 'HUB',
-      outputs: [{ internalType: 'address', name: '', type: 'address' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'to', type: 'address' },
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' }
-      ],
-      name: 'approve',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'address', name: 'owner', type: 'address' }],
-      name: 'balanceOf',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'burn',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' },
-        {
-          components: [
-            { internalType: 'uint8', name: 'v', type: 'uint8' },
-            { internalType: 'bytes32', name: 'r', type: 'bytes32' },
-            { internalType: 'bytes32', name: 's', type: 'bytes32' },
-            { internalType: 'uint256', name: 'deadline', type: 'uint256' }
-          ],
-          internalType: 'struct DataTypes.EIP712Signature',
-          name: 'sig',
-          type: 'tuple'
-        }
-      ],
-      name: 'burnWithSig',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'address', name: 'delegatee', type: 'address' }],
-      name: 'delegate',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'delegator', type: 'address' },
-        { internalType: 'address', name: 'delegatee', type: 'address' },
-        {
-          components: [
-            { internalType: 'uint8', name: 'v', type: 'uint8' },
-            { internalType: 'bytes32', name: 'r', type: 'bytes32' },
-            { internalType: 'bytes32', name: 's', type: 'bytes32' },
-            { internalType: 'uint256', name: 'deadline', type: 'uint256' }
-          ],
-          internalType: 'struct DataTypes.EIP712Signature',
-          name: 'sig',
-          type: 'tuple'
-        }
-      ],
-      name: 'delegateBySig',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'exists',
-      outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'getApproved',
-      outputs: [{ internalType: 'address', name: '', type: 'address' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'blockNumber', type: 'uint256' }],
-      name: 'getDelegatedSupplyByBlockNumber',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [],
-      name: 'getDomainSeparator',
-      outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'user', type: 'address' },
-        { internalType: 'uint256', name: 'blockNumber', type: 'uint256' }
-      ],
-      name: 'getPowerByBlockNumber',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'profileId', type: 'uint256' }],
-      name: 'initialize',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'owner', type: 'address' },
-        { internalType: 'address', name: 'operator', type: 'address' }
-      ],
-      name: 'isApprovedForAll',
-      outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'address', name: 'to', type: 'address' }],
-      name: 'mint',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'mintTimestampOf',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [],
-      name: 'name',
-      outputs: [{ internalType: 'string', name: '', type: 'string' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'ownerOf',
-      outputs: [{ internalType: 'address', name: '', type: 'address' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'spender', type: 'address' },
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' },
-        {
-          components: [
-            { internalType: 'uint8', name: 'v', type: 'uint8' },
-            { internalType: 'bytes32', name: 'r', type: 'bytes32' },
-            { internalType: 'bytes32', name: 's', type: 'bytes32' },
-            { internalType: 'uint256', name: 'deadline', type: 'uint256' }
-          ],
-          internalType: 'struct DataTypes.EIP712Signature',
-          name: 'sig',
-          type: 'tuple'
-        }
-      ],
-      name: 'permit',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'owner', type: 'address' },
-        { internalType: 'address', name: 'operator', type: 'address' },
-        { internalType: 'bool', name: 'approved', type: 'bool' },
-        {
-          components: [
-            { internalType: 'uint8', name: 'v', type: 'uint8' },
-            { internalType: 'bytes32', name: 'r', type: 'bytes32' },
-            { internalType: 'bytes32', name: 's', type: 'bytes32' },
-            { internalType: 'uint256', name: 'deadline', type: 'uint256' }
-          ],
-          internalType: 'struct DataTypes.EIP712Signature',
-          name: 'sig',
-          type: 'tuple'
-        }
-      ],
-      name: 'permitForAll',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'from', type: 'address' },
-        { internalType: 'address', name: 'to', type: 'address' },
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' }
-      ],
-      name: 'safeTransferFrom',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'from', type: 'address' },
-        { internalType: 'address', name: 'to', type: 'address' },
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' },
-        { internalType: 'bytes', name: '_data', type: 'bytes' }
-      ],
-      name: 'safeTransferFrom',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'operator', type: 'address' },
-        { internalType: 'bool', name: 'approved', type: 'bool' }
-      ],
-      name: 'setApprovalForAll',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'address', name: '', type: 'address' }],
-      name: 'sigNonces',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'bytes4', name: 'interfaceId', type: 'bytes4' }],
-      name: 'supportsInterface',
-      outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [],
-      name: 'symbol',
-      outputs: [{ internalType: 'string', name: '', type: 'string' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'index', type: 'uint256' }],
-      name: 'tokenByIndex',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'tokenDataOf',
-      outputs: [
-        {
-          components: [
-            { internalType: 'address', name: 'owner', type: 'address' },
-            {
-              internalType: 'uint96',
-              name: 'mintTimestamp',
-              type: 'uint96'
-            }
-          ],
-          internalType: 'struct IERC721Time.TokenData',
-          name: '',
-          type: 'tuple'
-        }
-      ],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'owner', type: 'address' },
-        { internalType: 'uint256', name: 'index', type: 'uint256' }
-      ],
-      name: 'tokenOfOwnerByIndex',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [{ internalType: 'uint256', name: 'tokenId', type: 'uint256' }],
-      name: 'tokenURI',
-      outputs: [{ internalType: 'string', name: '', type: 'string' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [],
-      name: 'totalSupply',
-      outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
-      stateMutability: 'view',
-      type: 'function'
-    },
-    {
-      inputs: [
-        { internalType: 'address', name: 'from', type: 'address' },
-        { internalType: 'address', name: 'to', type: 'address' },
-        { internalType: 'uint256', name: 'tokenId', type: 'uint256' }
-      ],
-      name: 'transferFrom',
-      outputs: [],
-      stateMutability: 'nonpayable',
-      type: 'function'
-    }
-  ];
