--- src/graphql/mutations/Broadcast.graphql
@@ -1,11 +1,0 @@
-mutation Broadcast($request: BroadcastRequest!) {
-  broadcast(request: $request) {
-    ... on RelayerResult {
-      txHash
-      txId
-    }
-    ... on RelayError {
-      reason
-    }
-  }
-}
