--- src/graphql/mutations/CreateFollowTypedData.graphql
@@ -1,26 +1,0 @@
-mutation CreateFollowTypedData($options: TypedDataOptions, $request: FollowRequest!) {
-  createFollowTypedData(options: $options, request: $request) {
-    id
-    expiresAt
-    typedData {
-      domain {
-        name
-        chainId
-        version
-        verifyingContract
-      }
-      types {
-        FollowWithSig {
-          name
-          type
-        }
-      }
-      value {
-        nonce
-        deadline
-        profileIds
-        datas
-      }
-    }
-  }
-}
