--- src/graphql/mutations/ProxyAction.graphql
@@ -1,3 +1,0 @@
-mutation ProxyAction($request: ProxyActionRequest!) {
-  proxyAction(request: $request)
-}
