--- src/graphql/profiles/UserProfiles.graphql
@@ -1,17 +1,0 @@
-query UserProfiles($ownedBy: [EthereumAddress!]) {
-  profiles(request: { ownedBy: $ownedBy }) {
-    items {
-      ...ProfileFields
-      stats {
-        totalFollowing
-      }
-      isDefault
-      dispatcher {
-        canUseRelay
-      }
-    }
-  }
-  userSigNonces {
-    lensHubOnChainSigNonce
-  }
-}
