+++ src/store/app.ts
@@ -1,4 +1,4 @@
-import { Profile, ReferenceModules } from "@/types/lens";
+import type { Profile } from "@/types/lens";
 import create from "zustand";
 import { persist } from "zustand/middleware";
 
@@ -34,21 +34,3 @@ export const useAppPersistStore = create(
     { name: "lenster.store" }
   )
 );
-
-interface ReferenceModuleState {
-  selectedReferenceModule: ReferenceModules;
-  setSelectedReferenceModule: (selectedModule: ReferenceModules) => void;
-  onlyFollowers: boolean;
-  setOnlyFollowers: (onlyFollowers: boolean) => void;
-  degreesOfSeparation: number;
-  setDegreesOfSeparation: (degreesOfSeparation: number) => void;
-}
-
-export const useReferenceModuleStore = create<ReferenceModuleState>((set) => ({
-  selectedReferenceModule: ReferenceModules.DegreesOfSeparationReferenceModule,
-  setSelectedReferenceModule: (selectedReferenceModule) => set(() => ({ selectedReferenceModule })),
-  onlyFollowers: false,
-  setOnlyFollowers: (onlyFollowers) => set(() => ({ onlyFollowers })),
-  degreesOfSeparation: 2,
-  setDegreesOfSeparation: (degreesOfSeparation) => set(() => ({ degreesOfSeparation }))
-}));
